requirements.txt
easydict==1.9
ipdb==0.13.3
lmdb==1.0.0
pyflakes==2.2.0
PyYAML==5.3.1
tqdm==4.46.0
tensorboard==2.3.0
scikit-learn==0.24.0
omegaconf
pickle5
einops
imageio
torchaudio==0.13.0
ftfy
regex
datadings